requirements.txt
Flask
flask-cors
Pillow
torch
skimage
